requirements.txt
# Core Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
openpyxl>=3.1.0

# Configuration Management
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI/ML
openai>=1.6.0

# Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0

# Utilities
tqdm>=4.66.0

# Development
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
